dev(SessionProvider): move logic from controller to service
